requirements/production.txt
# Pro-tip: Try not to put anything here. Avoid dependencies in
#	production that aren't in development.
-r base.txt

# Python-PostgreSQL Database Adapter
# Assuming Windows is used locally, and *nix -- in production.
# ------------------------------------------------------------
psycopg2==2.7.3.2

# WSGI Handler
# ------------------------------------------------
gevent==1.2.2
gunicorn==19.7.1

# Static and Media Storage
# ------------------------------------------------
boto3==1.4.7
django-storages==1.6.5
Collectfast==0.5.2

# Email backends for Mailgun, Postmark, SendGrid and more
# -------------------------------------------------------
django-anymail==1.2




